Models/ full repo

+++ GestionDesRessourcesMaterielles/Models/Departement.cs
@@ -8,7 +8,7 @@ namespace GestionDesRessourcesMaterielles.Models
         [Key]
         public int DepartmentId { get; set; }
         public string Name { get; set; }
-        public float Budget { get; set; }
+        public double Budget { get; set; }
 
     }
 }
